requirements.txt
fastmcp
pyyaml
pydantic-settings>=2.0.0
pytest>=8.0.0
ruff>=0.4.0
black>=24.0.0
# OpenTelemetry SDK and exporters (latest stable for full compatibility)
opentelemetry-api==1.33.1
opentelemetry-sdk==1.33.1
opentelemetry-exporter-otlp==1.33.1
opentelemetry-exporter-prometheus==0.54b1
# Add other OTEL exporters as needed
